timestamp: extract timestamps to their own module
