+++ src/transaction.rs
@@ -1,61 +1,8 @@
-use std::time::SystemTime;
-
 use super::application::Application;
 use super::user::User;
 use super::usage::Usage;
 use super::ToParams;
-
-#[derive(Clone, Debug)]
-pub struct Timestamp {
-    ts: String,
-}
-
-impl Timestamp {
-    pub fn none() -> Option<&'static Timestamp> {
-        None
-    }
-
-    pub fn new(ts: u64) -> Self {
-        Self { ts: ts.to_string() }
-    }
-
-    pub fn into_inner(self) -> String {
-        self.ts
-    }
-
-    pub fn as_str(&self) -> &str {
-        self.ts.as_str()
-    }
-}
-
-impl<'ts> AsRef<str> for Timestamp {
-    fn as_ref(&self) -> &str {
-        self.ts.as_str()
-    }
-}
-
-impl From<u64> for Timestamp {
-    fn from(ts: u64) -> Self {
-        Self::new(ts)
-    }
-}
-
-impl From<SystemTime> for Timestamp {
-    fn from(st: SystemTime) -> Self {
-        let ts = match st.duration_since(SystemTime::UNIX_EPOCH) {
-            Ok(ts) => ts.as_secs(),
-            _ => 0,
-        };
-
-        Self::new(ts)
-    }
-}
-
-impl Default for Timestamp {
-    fn default() -> Self {
-        Timestamp::from(SystemTime::now())
-    }
-}
+use super::timestamp::Timestamp;
 
 #[derive(Copy, Clone, Debug)]
 pub struct Transaction<'app, 'user, 'usage, 'ts> {
